requirements.txt
python-telegram-bot==20.6
requests==2.31.0
google-generativeai==0.3.1
python-dotenv==1.0.0